# dify/openapi_tools_v5.yaml
openapi: "3.0.0"
info:
  title: LUTAGU Tools
  description: Transit tools for Dify Agent
  version: "1.0.0"
servers:
  - url: https://lutagu.vercel.app
paths:
...
  /api/odpt/route:
    get:
      operationId: search_route
      summary: Search route
      description: Find route between stations
      parameters:
        - name: from
          in: query
          required: true
          description: Origin station
          schema:
            type: string
        - name: to
          in: query
          required: true
          description: Destination station
          schema:
            type: string
        - name: locale
          in: query
          required: false
          description: Response language
          schema:
            type: string
            enum: [ja, en, zh-TW]
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
...